ai-projects/projects/influencer-shortlist-agent/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langfuse>=4.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
qdrant-client>=1.7.0
cohere>=5.0.0
redis>=5.0.0
sqlglot>=23.0.0
sentence-transformers>=2.2.0
fastembed>=0.3.0
importlib_metadata>=6.0,<8.8.0
